2024/10/8
Rewrite the method for solving the Jacobi matrix in the solver from using numerical differentiation to using an analytical expression. In this version of the Newton iteration solver, no loops were used except for the outermost while loop, effectively avoiding the weakness of Python's slow loop execution and significantly improving the solving speed.